Language Learning Algorithm
Began making functions to implement the language learning algorithm that fetches cards from the database in a phased manner. At the moment, the functions are standalone and have not been implemented into the front end. The 'Algorithm' folder will contain copies of the front end, in future versions, where I will test the algorithm on the front end.